Refactor main.rb file using SOLID
